Rename keys of app's global state object

+++ models/settings.js
@@ -2,8 +2,8 @@ const mongoose = require('mongoose')
 
 // Create model and schema. Export model.
 var Setting = mongoose.model('settings', new mongoose.Schema({
-    dlMode: {type: String, default: 'browser'},
-    dlType: {type: String, default: 'video'}
+    mode: {type: String, default: 'browser'},
+    format: {type: String, default: 'video'}
 }))
 exports.Setting = Setting
 
@@ -12,21 +12,20 @@ async function main() {
 
     // Should only be 1 document in collection. Recreate if more than 1
     const count = await Setting.countDocuments({})
-    
-    if (count > 1) {
-        Setting.collection.drop()
-    }
+    if (count > 1) {Setting.collection.drop()}
 
     // Find first document where correct keys exist
     const settings = await Setting.findOne({
-        dlMode: {$exists: true},
-        dlType: {$exists: true}
+        mode: {$exists: true},
+        format: {$exists: true}
     })
 
     if (settings) {
         // Settings exist
     } else {
-        // Settings do not exist. Create with default values.
+        // Settings do not exist or do not have correct key/values. Create with default values.
+        Setting.collection.drop().catch(e => {})
+
         const document = new Setting()
         document.save().catch(e => console.error(e))
     }
